backend/requirements.txt
# Core Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart

# Data Models & Validation
pydantic>=2.10.0
pydantic-settings>=2.2.0

# HTTP Client
httpx>=0.27.0
requests

# WebSocket
websockets>=12.0

# Database & Caching
redis==5.0.2
firebase_admin

# Resume Parsing
PyPDF2==3.0.1
python-docx==0.8.11

# AI/ML Services
google-generativeai
numpy<2.0.0

# Speech Services
deepgram-sdk>=3.2.7,<4.0.0
elevenlabs>=1.2.2,<2.0.0

# Code Execution
# (Judge0 via HTTP API)

# Utilities
python-dotenv==1.0.1
aiohttp
aiofiles
tenacity